Linux cleanup
With this change, the linux binaries built with rad no longer
have a strong dependency on the crt. One of the main issues
is that the stack might not be aligned on the start
function, that is fixed now.
Overall there are still many dependencies due to how Basic.so
pulls in, due to malloc and free.