LLVM Optimizations
Now we can optimize our code. Rad supports the `-opt` Option. There is code in case we want to go full -O1, -O2, etc.
To do this, we needed to clean up how we handle return statements, there is now tracking for which return statements are at
function level, and we also discard statements after the return statement.